% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_ghost.R
\name{ghost_cluster}
\alias{ghost_cluster}
\title{Helper function to create potential ghost hostel clusters}
\usage{
ghost_cluster(property, distance, min_listings)
}
\arguments{
\item{property}{A data frame of STR listings with sf or sp point geometries in
a projected coordinate system.}

\item{distance}{A numeric scalar. The radius (in the units of the CRS) of the
buffer which will be drawn around points to determine possible ghost hostel
locations.}

\item{min_listings}{A numeric scalar. The minimum number of listings to
be considered a ghost hostel.}
}
\value{
The `property` table, rearranged with one row per cluster and with a
new `predicates` field.
}
\description{
Helper function to create potential ghost hostel clusters
}
