% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_ghost_old_helpers.R
\name{ghost_combine}
\alias{ghost_combine}
\title{Helper function to create potential intersection combinations}
\usage{
ghost_combine(buffers, predicates, n)
}
\arguments{
\item{buffers}{A data frame of buffers from property$data.}

\item{predicates}{The property$data field generated from \code{st_intersects}.}

\item{n}{A numeric scalar. The number of points to attempt to find a set of
combinations for}
}
\value{
A matrix of possible intersection combinations.
}
\description{
Helper function to create potential intersection combinations
}
