% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_ghost.R
\name{ghost_empty}
\alias{ghost_empty}
\title{Helper function to output an empty ghost hostel table}
\usage{
ghost_empty(crs_property)
}
\arguments{
\item{crs_property}{The CRS of the property table.}
}
\description{
Helper function to output an empty ghost hostel table
}
