% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_ghost.R
\name{ghost_identify_leftovers}
\alias{ghost_identify_leftovers}
\title{Helper function to identify leftover ghost hostel candidates}
\usage{
ghost_identify_leftovers(property, min_listings)
}
\arguments{
\item{property}{An sf data frame of STR listings nested by cluster.}

\item{min_listings}{A numeric scalar. The minimum number of listings to
be considered a ghost hostel.}
}
\value{
The output will be a set of new candidate points.
}
\description{
Helper function to identify leftover ghost hostel candidates
}
