% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_ghost.R
\name{ghost_intersect}
\alias{ghost_intersect}
\title{Helper function to find ghost hotel locations}
\usage{
ghost_intersect(points, property_ID, host_ID, distance, min_listings)
}
\arguments{
\item{points}{An sf data frame of STR listings nested by cluster.}

\item{property_ID}{The name of a character or numeric variable in the points
object which uniquely identifies STR listings.}

\item{host_ID}{The name of a character or numeric variable in the points
object which uniquely identifies STR hosts.}

\item{distance}{A numeric scalar. The radius (in the units of the CRS) of the
buffer which will be drawn around points to determine possible ghost hotel
locations.}

\item{min_listings}{A numeric scalar. The minimum number of listings to
be considered a ghost hotel.}
}
\value{
The output will be `points`, trimmed to valid ghost hotel locations
  and with additional fields added.
}
\description{
\code{ghost_intersect} finds ghost hotel locations from a set of clusters.
}
\details{
A function for finding ghost hotel locations from a set of point clusters.
}
