% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_ghost_2.R
\name{ghost_intersect_2}
\alias{ghost_intersect_2}
\title{Helper function to find ghost hostel locations}
\usage{
ghost_intersect_2(points, distance, min_listings)
}
\arguments{
\item{points}{An sf data frame of STR listings nested by cluster.}

\item{distance}{A numeric scalar. The radius (in the units of the CRS) of the
buffer which will be drawn around points to determine possible ghost hostel
locations.}

\item{min_listings}{A numeric scalar. The minimum number of listings to
be considered a ghost hostel.}
}
\value{
The output will be `points`, trimmed to valid ghost hostel locations
  and with additional fields added.
}
\description{
\code{ghost_intersect} finds ghost hostel locations from a set of clusters.
}
\details{
A function for finding ghost hostel locations from a set of point clusters.
}
