% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_ghost.R
\name{ghost_intersect_leftovers}
\alias{ghost_intersect_leftovers}
\title{Helper function to rerun analysis on leftover ghost hotel candidates}
\usage{
ghost_intersect_leftovers(points, property_ID, host_ID, distance,
  min_listings)
}
\arguments{
\item{points}{An sf data frame of STR listings nested by cluster.}

\item{property_ID}{The name of a character or numeric variable in the points
object which uniquely identifies STR listings.}

\item{host_ID}{The name of a character or numeric variable in the points
object which uniquely identifies STR hosts.}

\item{distance}{A numeric scalar. The radius (in the units of the CRS) of the
buffer which will be drawn around points to determine possible ghost hotel
locations.}

\item{min_listings}{A numeric scalar. The minimum number of listings to
be considered a ghost hotel.}
}
\value{
The output will be the `points` file with additional ghost hotels
  added.
}
\description{
\code{ghost_intersect_leftovers} reruns the ghost hotel analysis on leftover
 candidates.
}
\details{
A function for rerunning the ghost hotel analysis on leftover candidates
after the main analysis has been run.
}
