% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_ghost_2.R
\name{ghost_intersect_leftovers_2}
\alias{ghost_intersect_leftovers_2}
\title{Helper function to rerun analysis on leftover ghost hostel candidates}
\usage{
ghost_intersect_leftovers_2(points, distance, min_listings)
}
\arguments{
\item{points}{An sf data frame of STR listings nested by cluster.}

\item{distance}{A numeric scalar. The radius (in the units of the CRS) of the
buffer which will be drawn around points to determine possible ghost hostel
locations.}

\item{min_listings}{A numeric scalar. The minimum number of listings to
be considered a ghost hostel.}
}
\value{
The output will be the `points` file with additional ghost hostels
  added.
}
\description{
\code{ghost_intersect_leftovers} reruns the ghost hostel analysis on leftover
 candidates.
}
\details{
A function for rerunning the ghost hostel analysis on leftover candidates
after the main analysis has been run.
}
