% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_ghost.R, R/strr_ghost_2.R
\name{ghost_intersect_with_done}
\alias{ghost_intersect_with_done}
\title{Helper function to intersect two buffers}
\usage{
ghost_intersect_with_done(x, y)

ghost_intersect_with_done(x, y)
}
\arguments{
\item{x}{A buffer.}

\item{y}{A buffer.}
}
\value{
The output will be an intersect polygon.

The output will be an intersect polygon.
}
\description{
\code{ghost_intersect_with_done} takes the intersection of two buffers.

\code{ghost_intersect_with_done} takes the intersection of two buffers.
}
\details{
A function for intersecting two buffers with done(), for use in reduce().

A function for intersecting two buffers with done(), for use in reduce().
}
