% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_ghost_2.R
\name{ghost_stepwise_intersect_2}
\alias{ghost_stepwise_intersect_2}
\title{Helper function to intersect a data frame of buffers}
\usage{
ghost_stepwise_intersect_2(buffers, min_listings)
}
\arguments{
\item{buffers}{A data frame of buffers from points$data.}

\item{min_listings}{A numeric scalar. The minimum number of listings to
be considered a ghost hostel.}
}
\value{
The output will be an intersect polygon.
}
\description{
\code{ghost_stepwise_intersect} applies st_intersection in stepwise fashion
to a data frame of buffers.
}
\details{
A function for intersecting a data frame of buffers in stepwise fashion,
including a number of optimizations for performance.
}
