% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_helpers.R
\name{handler_strr}
\alias{handler_strr}
\title{Helper function to set a progress reporting strategy}
\usage{
handler_strr(message)
}
\arguments{
\item{message}{A character string describing the task being iterated.}
}
\value{
No visible output.
}
\description{
\code{handler_strr} sets a progress reporting strategy.
}
