% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_helpers.R
\name{helper_check_daily}
\alias{helper_check_daily}
\title{Helper function to check an strr_daily data frame}
\usage{
helper_check_daily(...)
}
\arguments{
\item{...}{Field names to be passed as quoted symbol names (with
`rlang::ensym`).}
}
\value{
An error if the check fails.
}
\description{
\code{helper_check_daily} checks a data frame and verifies that is strr_daily
class, and optionally verifies the presence of specified fields in the data
frame.
}
