% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_helpers.R
\name{helper_check_data}
\alias{helper_check_data}
\title{Helper function to check a data frame}
\usage{
helper_check_data()
}
\value{
A logical scalar "daily" indicating if the table is strr_daily or
strr_host, or an error if the check fails.
}
\description{
\code{helper_check_data} checks a data frame and verifies that is either
strr_daily or strr_host.
}
