% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_helpers.R
\name{helper_check_host}
\alias{helper_check_host}
\title{Helper function to check a host data frame}
\usage{
helper_check_host()
}
\value{
An error if the check fails.
}
\description{
\code{helper_check_host} checks a data frame and verifies that is strr_host
class.
}
