% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_helpers.R
\name{helper_check_property}
\alias{helper_check_property}
\title{Helper function to check a property data frame}
\usage{
helper_check_property(...)
}
\arguments{
\item{...}{Field names to be passed as quoted symbol names (with
`rlang::ensym`).}
}
\value{
An error if the check fails.
}
\description{
\code{helper_check_property} checks a data frame and verifies that is
strr_property class.
}
