% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_helpers.R
\name{helper_check_quiet}
\alias{helper_check_quiet}
\title{Helper function to check a `quiet` argument}
\usage{
helper_check_quiet()
}
\value{
An error if the check fails.
}
\description{
\code{helper_check_quiet} checks a `quiet` argument and verifies that it is
a logical scalar.
}
