% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_compress.R
\name{helper_compress_daily}
\alias{helper_compress_daily}
\title{Helper function to compress daily file}
\usage{
helper_compress_daily(data)
}
\arguments{
\item{data}{The processed daily table generated through the strr_compress
function.}
}
\value{
The output will be a compressed daily table.
}
\description{
\code{helper_compress_daily} takes a processed `daily` table and
generates a compressed version.
}
\details{
A helper function for compressing the processed monthly `daily` table.
}
