% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_expand.R
\name{helper_expand}
\alias{helper_expand}
\title{Helper function to expand compressed STR tables}
\usage{
helper_expand(x)
}
\arguments{
\item{x}{Argument passed from the main function.}
}
\description{
Helper function to expand compressed STR tables
}
