% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_raffle_helpers.R
\name{helper_integrate}
\alias{helper_integrate}
\title{Helper function to integrate raffle_pdf_* over intersect polygons}
\usage{
helper_integrate(x, y, pdf)
}
\arguments{
\item{x, y, pdf}{Arguments passed along from main function.}
}
\description{
Helper function to integrate raffle_pdf_* over intersect polygons
}
