% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_raffle_helpers.R
\name{helper_intersect}
\alias{helper_intersect}
\title{Helper function to intersect property and polys}
\usage{
helper_intersect(x, polys, distance)
}
\arguments{
\item{x, polys, distance}{Arguments passed along from the main function.}
}
\description{
Helper function to intersect property and polys
}
