% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_helpers.R
\name{helper_message}
\alias{helper_message}
\title{Helper function to display a message}
\usage{
helper_message(..., .type = "main", .quiet = NULL)
}
\arguments{
\item{...}{Character strings to be displayed.}

\item{.type}{One of c("open", "close", "main", "final"). "Open" prints a
temporary message in grey italics with no timestamp. "Close" is designed to
be called after "open", since it overwrites the previous line with a message
in grey with a timestamp in cyan. "Main" is the same as "close" but does not
override the previous line. "Final" is the same as "main" but appends an
additional line in cyan bold which states the total time.}

\item{.quiet}{The name of the argument in the calling function specifying
whether messages should be displayed.}
}
\value{
A status message.
}
\description{
\code{helper_message} produces a formatted progress message.
}
