% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_helpers.R
\name{helper_plan}
\alias{helper_plan}
\title{Helper function to characterize future plan}
\usage{
helper_plan()
}
\value{
A character string reporting the current plan.
}
\description{
\code{helper_plan} parses the current \code{future::plan} for display in
progress messages.
}
