% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_expand.R
\name{helper_prepare_expand}
\alias{helper_prepare_expand}
\title{Helper function to prepare to expand compressed STR tables}
\usage{
helper_prepare_expand(data, daily_flag)
}
\arguments{
\item{data, daily_flag}{Arguments passed from the main function.}
}
\description{
Helper function to prepare to expand compressed STR tables
}
