% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_raffle_helpers.R
\name{helper_prepare_grid}
\alias{helper_prepare_grid}
\title{Helper function to prepare grid}
\usage{
helper_prepare_grid(property, polys, distance)
}
\arguments{
\item{property, polys, distance}{Arguments passed along from the main
function.}
}
\value{
The output will be a list with the `property` table, the grid
geometry, and the number of iterations.
}
\description{
Helper function to prepare grid
}
