% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_raffle_helpers.R
\name{helper_prepare_integrate}
\alias{helper_prepare_integrate}
\title{Helper function to prepare to integrate intersects}
\usage{
helper_prepare_integrate(intersects)
}
\arguments{
\item{intersects}{Argument passed along from the main function.}
}
\description{
Helper function to prepare to integrate intersects
}
