% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_raffle_helpers.R
\name{helper_prepare_intersect}
\alias{helper_prepare_intersect}
\title{Helper function to prepare to intersect property and polys}
\usage{
helper_prepare_intersect(property)
}
\arguments{
\item{property}{Argument passed along from the main function.}
}
\description{
Helper function to prepare to intersect property and polys
}
