% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_raffle_helpers.R
\name{helper_prepare_polys}
\alias{helper_prepare_polys}
\title{Helper function to prepare polys}
\usage{
helper_prepare_polys(property, polys, poly_ID_flag)
}
\arguments{
\item{property, polys, poly_ID_flag}{Arguments passed along from the main
function.}
}
\value{
The output will be the `polys` table.
}
\description{
Helper function to prepare polys
}
