% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_raffle_helpers.R
\name{helper_prepare_property}
\alias{helper_prepare_property}
\title{Helper function to prepare property}
\usage{
helper_prepare_property(property, polys, poly_ID_flag)
}
\arguments{
\item{property, polys, poly_ID_flag}{Arguments passed along from the main
function.}
}
\value{
The output will be the `property` table.
}
\description{
Helper function to prepare property
}
