% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_raffle_helpers.R
\name{helper_process_intersects}
\alias{helper_process_intersects}
\title{Helper function to process intersections of property and polys}
\usage{
helper_process_intersects(intersects, empty)
}
\arguments{
\item{intersects, empty}{Arguments passed along from the main function.}
}
\description{
Helper function to process intersections of property and polys
}
