% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_raffle_helpers.R
\name{helper_process_results}
\alias{helper_process_results}
\title{Helper function to process raffle results}
\usage{
helper_process_results(
  property,
  result,
  diagnostic,
  poly_ID_flag,
  poly_ID_name
)
}
\arguments{
\item{property, result, diagnostic, poly_ID_flag, poly_ID_name}{Arguments passed
along from the main function.}
}
\value{
The output will be the `property` table.
}
\description{
Helper function to process raffle results
}
