% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_helpers.R
\name{helper_progress}
\alias{helper_progress}
\title{Helper function to display furrr progress bars}
\usage{
helper_progress(quiet)
}
\arguments{
\item{quiet}{A logical scalar. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
A logical scalar. Should the progress bar be displayed or not?
}
\description{
\code{helper_progress} decides whether to display progress bars in furrr
functions.
}
\details{
A helper function for deciding whether to display a progress bar when a
function is called from \code{furrr}, based on the supplied `quiet` argument
and the status of the future plan.
}
