% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_helpers.R
\name{helper_progress_message}
\alias{helper_progress_message}
\title{Helper function to display a progress message}
\usage{
helper_progress_message(..., .quiet = NULL, .final = FALSE)
}
\arguments{
\item{...}{Character strings to be displayed. The strings can include
code for evaluation via \code{glue::glue} inside `{}`.}

\item{.quiet}{The name of the argument in the calling function specifying
whether messages should be displayed.}

\item{.final}{A logical scalar. Is this the final progress message, to be
formatted in cyan and bold and display the total time?}
}
\value{
A status message.
}
\description{
\code{helper_progress_message} produces a formatted progress message,
optionally with the current time.
}
