% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_raffle.R
\name{helper_raffle_integrate}
\alias{helper_raffle_integrate}
\title{Helper function to integrate raffle_pdf_* over intersect polygons}
\usage{
helper_raffle_integrate(intersects)
}
\arguments{
\item{intersects}{A data.table with an sf geometry column.}
}
\value{
The output will be the input data.table with a `probability` field
added.
}
\description{
\code{helper_raffle_integrate} takes raffle_pdf_* and integrates it over the
intersect polygons using polyCub.
}
\details{
A function for integrating raffle_pdf_* over the intersect polygons using
polyCub. In the future, other PDFs may be added.
}
