% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_helpers.R
\name{helper_table_split}
\alias{helper_table_split}
\title{Helper function to split data for parallel processing}
\usage{
helper_table_split(data_list, multiplier = 4)
}
\arguments{
\item{data_list}{A list of data elements to be resized}

\item{multiplier}{An integer scalar. What multiple of the number of available
processes should the data be combined into?}
}
\value{
A list of data elements.
}
\description{
\code{helper_table_split} takes a split table and combines it into an
optimal number of elements for parallel processing.
}
