% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_helpers.R
\name{helper_test_field}
\alias{helper_test_field}
\title{Helper function to test for a field}
\usage{
helper_test_field(data, field, arg_name)
}
\arguments{
\item{data}{The table to be checked.}

\item{field}{The field to be checked.}

\item{arg_name}{A character string supplying the underlying argument name for
the field being checked.}
}
\value{
An error if the table is not present.
}
\description{
\code{helper_test_field} tests for the presence of a given field in the input
table.
}
