% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_helpers.R
\name{helper_total_time}
\alias{helper_total_time}
\title{Helper function to calculate the total function time}
\usage{
helper_total_time(time_1)
}
\arguments{
\item{time_1}{The time when the function began.}
}
\value{
A character string with the total function time.
}
\description{
\code{helper_total_time} produces a character string with the total function
time.
}
