% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_raffle.R
\name{raffle_integrate}
\alias{raffle_integrate}
\title{Helper function to integrate raffle_PDF over intersect polygons}
\usage{
raffle_integrate(intersects)
}
\arguments{
\item{intersects}{An sf data frame.}
}
\value{
The output will be the input data frame with a `probability` field
added.
}
\description{
\code{raffle_integrate} takes raffle_PDF and integrates it over the intersect
polygons using polyCub.
}
\details{
A function for integrating raffle_PDF (Airbnb's spatial obfuscation of
listing locations) over the intersect polygons using polyCub. In the future,
other PDFs may be added.
}
