% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_raffle.R
\name{raffle_pdf}
\alias{raffle_pdf}
\title{Helper function specifying the probability density function}
\usage{
raffle_pdf(x)
}
\arguments{
\item{x}{An st_coordinates object, converted to sp.}
}
\value{
The output will be a numerical vector of probabilities.
}
\description{
\code{raffle_pdf} specifies the probability density function which Airbnb
uses to obfuscate listing locations.
}
\details{
A function which specifies the probability density function which Airbnb
uses to obfuscate listing locations (a radially symmetric normal distribution
with mean 100 m and standard deviation 50 m, truncated at r = 0 and r = 200).
In the future, other PDFs may be added.
}
