% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_ghost_helpers.R
\name{reduce}
\alias{reduce}
\title{Helper function to reduce predicate lists to minimal overlapping sublists}
\usage{
reduce(x)
}
\arguments{
\item{x}{A predicate list, or other list of numeric vectors.}
}
\value{
A list whose elements are the minimal overlapping sublists of 
elements of x.
}
\description{
Helper function to reduce predicate lists to minimal overlapping sublists
}
