% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_as_sf.R
\name{strr_as_sf}
\alias{strr_as_sf}
\title{Function to convert property tables to sf objects}
\usage{
strr_as_sf(.data, CRS = 4326, longitude = "longitude", latitude = "latitude")
}
\arguments{
\item{.data}{A property table in standard UPGo format.}

\item{CRS}{The EPSG code or proj4string character vector specifying a
coordinate reference system into which to transform the table. The default
is unprojected WGS 84 (EPSG 4326).}

\item{longitude}{A character string naming the field supplying the
longitude values of points.}

\item{latitude}{A character string naming the field supplying the
latitude values of points.}
}
\value{
An sf table with the longitude and latitude fields converted to a
geometry field and all other fields returned unaltered.
}
\description{
\code{strr_as_sf} is a convenience wrapper around \code{sf::st_as_sf} for STR
property tables.
}
\details{
A function for converting a raw property table into an sf object with point
geometries. The function makes use of a pair of fields specifying longitude
and latitude, and has the option to specify a destination CRS.
}
