% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_compress_daily.R
\name{strr_compress_daily}
\alias{strr_compress_daily}
\title{Function to compress raw daily STR tables into UPGo DB format}
\usage{
strr_compress_daily(daily, output_date = NULL, cores = 1,
  quiet = FALSE)
}
\arguments{
\item{daily}{A daily table in the raw AirDNA format.}

\item{output_date}{A character string of the format "YYYY-MM" identifying the
year and month of the daily table. If a value is supplied to this argument,
the results of the function will be written to disk in a subfolder
"/output/YYYY/". If the argument is left NULL (its default), nothing will be
written to disk.}

\item{cores}{A positive integer scalar. How many processing cores should be
used to perform the computationally intensive compression step?}

\item{quiet}{A logical vector. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
The output will be a list with three elements: 1) the compressed
daily table, ready for upload to a remote database; 2) an error table
identifying corrupt or otherwise invalid row entries; 3) a missing_rows
table identifying property_IDs with missing dates in between their first and
last date entries, and therefore potentially missing data.
}
\description{
\code{strr_compress_daily} takes raw monthyl daily tables from AirDNA and
compresses them into the UPGo database storage format.
}
\details{
A function for compressing the daily activity tables supplied each month from
AirDNA into a more storage-efficient one-activity-block-per-row format. The
function also produces error files which identify possible corrupt or missing
lines in the input file.
}
