% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_compress.R
\name{strr_compress_helper_ML}
\alias{strr_compress_helper_ML}
\title{Helper function to compress daily file}
\usage{
strr_compress_helper_ML(.data, dates)
}
\arguments{
\item{.data}{The processed ML table generated through the strr_compress
function.}

\item{dates}{A logical scalar. Were `year` and `month` fields created in the
input table?}
}
\value{
The output will be a compressed ML table.
}
\description{
\code{strr_compress_helper_ML} takes a processed ML table and generates a
compressed version.
}
\details{
A helper function for compressing the processed ML summary table.
}
