% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_expand_daily.R
\name{strr_expand_daily}
\alias{strr_expand_daily}
\title{Function to expand compressed daily STR tables}
\usage{
strr_expand_daily(daily, start = NULL, end = NULL, cores = 1)
}
\arguments{
\item{daily}{A daily table in the compressed UPGo DB format (e.g. created
by running \code{\link{strr_compress_daily}} on a raw daily table from
AirDNA).}

\item{start}{A character string of format YYYY-MM-DD indicating the
first date to be provided in the output table. If NULL (default), the
earliest date present in the data will be used.}

\item{end}{A character string of format YYYY-MM-DD indicating the
last date to be provided in the output table. If NULL (default), the
latest date present in the data will be used.}

\item{cores}{A positive integer scalar. How many processing cores should be
used to perform the computationally intensive numeric integration step?}
}
\value{
A table of daily STR activity with one row per date and all other
fields returned unaltered.
}
\description{
\code{strr_expand_daily} takes STR daily file compressed in the UPGo DB
format and expands it to a one-row-per-date format.
}
\details{
A function for expanding compressed daily activity tables from AirDNA. The
function will also optionally truncated the table by a supplied date range.
}
