% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_expand.R
\name{strr_expand_helper}
\alias{strr_expand_helper}
\title{Helper function to expand compressed STR tables}
\usage{
strr_expand_helper(data, daily_flag, quiet)
}
\arguments{
\item{data}{A table in compressed UPGo DB format (e.g. created by running
\code{\link{strr_compress}}). Currently daily activity table and host tables
are recognized.}

\item{daily_flag}{A logical scalar. Is the input table a daily table (TRUE)
or a host table (FALSE)?}

\item{quiet}{A logical scalar. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
A table with one row per date and all other fields returned
unaltered.
}
\description{
\code{strr_expand_helper} performs the daily/host table expansion within
\code{\link{strr_expand}}.
#'
}
