% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_FREH.R
\name{strr_FREH}
\alias{strr_FREH}
\title{Function to identify frequently rented entire-home (FREH) listings}
\usage{
strr_FREH(
  daily,
  start_date,
  end_date,
  property_ID = property_ID,
  date = date,
  status = status,
  status_types = c("R", "A"),
  listing_type = listing_type,
  entire_home = "Entire home/apt",
  n_days = 365,
  r_cut = 90,
  ar_cut = 183,
  quiet = FALSE
)
}
\arguments{
\item{daily}{A data frame of daily STR activity in standard UPGo format.}

\item{start_date}{A character string of format YYYY-MM-DD indicating the
first date for which to return output. If NULL (default), all dates will
be used.}

\item{end_date}{A character string of format YYYY-MM-DD indicating the last
date for which to run the analysis.}

\item{property_ID}{The name of a character or numeric variable in the `daily`
table which uniquely identifies STR listings.}

\item{date}{The name of a date variable in the `daily` table.}

\item{status}{The name of a character variable in the `daily` table which
identifies the activity status of a listing on a give date.}

\item{status_types}{A two-length character vector which identifies the
values in the `status` variable indicating "reserved" and "available" status
respectively. The default value is \code{c("R", "A")}.}

\item{listing_type}{The name of a character variable in the `daily`
table which identifies entire-home listings. Set this argument to FALSE
to use all listings in the `daily` table.}

\item{entire_home}{A character string which identifies the value of the
`listing_type` variable to be used to find entire-home listings. This field
is ignored if `listing_type` is FALSE.}

\item{n_days}{An integer scalar which determines how many days should be used
to evaluate each listing's activity status. The default is 365 days (one
year).}

\item{r_cut}{TKTK}

\item{ar_cut}{TKTK}

\item{quiet}{A logical vector. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
The output will be a tidy data frame of identified FREH listings,
  organized with the following fields: `property_ID` (or whatever name was
  passed to the property_ID argument): A character vector with the ID code of
  the listings. `date`: The date for which the FREH status is being reported.
  `FREH`: A logical scalar indicating whether, on a given date, the given
  listing exceeded the `r_cut` and `ar_cut` thresholds over the number of
  days specified by `n_days`.
}
\description{
\code{strr_FREH} takes a table of daily STR activity and identifies listings
which met a given standard of availability and activity over a specified
time period.
}
\details{
TKTK
}
