% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_housing.R
\name{strr_housing}
\alias{strr_housing}
\title{Function to identify STRs located in housing}
\usage{
strr_housing(property, property_type = property_type)
}
\arguments{
\item{property}{An STR property file.}

\item{property_type}{The name of a field identifying the type of property a
listing is located in.}
}
\value{
The output will be the input `property` table, but with an additional
`housing` field added, with TRUE or FALSE values indicating if a given
listing is located in housing (TRUE) or in a dedicated tourist accommodation
facility such as a hotel or B&B (FALSE).
}
\description{
\code{strr_housing} takes an STR property file and adds a `housing` field.
}
\details{
A function for adding a `housing` field to an STR property file. The function
reads the `property_type` fields and categorizes the values into either
housing or not housing.
}
