% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_multi.R
\name{strr_multi}
\alias{strr_multi}
\title{Function to identify STR multilistings}
\usage{
strr_multi(
  daily,
  host,
  thresholds = c(EH = 2L, PR = 3L, SR = NA, HR = NA),
  combine_housing = TRUE,
  field_name = multi,
  quiet = FALSE
)
}
\arguments{
\item{daily}{A data frame of daily STR activity in standard UPGo format.}

\item{host}{A data frame of STR host activity in standard UPGo format.}

\item{thresholds}{An integer vector indicating the thresholds to establish
multilisting status for each listing type. If the vector is unnamed, the
options will be read in the order "Entire home/apt", "Private room",
"Shared room", "Hotel room". If the vector is named, using the full listing
type names or acronyms ("EH", "PR", "SR", "HR"), the values can be presented
in any order. An NA or 0 value means that multilisting calculations will not
be performed for that listing type.}

\item{combine_housing}{A logical scalar. Should housing and non-housing
listings be combined for the purposes of establishing multilisting status
(default)?}

\item{field_name}{A symbol or character string to name the new logical field
identifying multilisting status in the output table. The default is
`multi`.}

\item{quiet}{A logical scalar. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
The output will be the `daily` input table with one additional
logical field (with name taken from the `field_name` argument) indicating
multilisting status.
}
\description{
\code{strr_multi} takes a daily table and a host table and identifies which
listings in the daily table are multilistings on each date.
}
\details{
TKTK explanation
}
