% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_process_daily_2.R
\name{strr_process_daily_helper}
\alias{strr_process_daily_helper}
\title{Helper function to process daily file}
\usage{
strr_process_daily_helper(daily)
}
\arguments{
\item{daily}{TKTK}
}
\value{
TKTK
}
\description{
\code{strr_process_daily_helper} takes a processed host table and generates a
compressed version.
}
\details{
TKTK
}
