% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_process_multi.R
\name{strr_process_multi}
\alias{strr_process_multi}
\title{Function to create daily multilisting tables}
\usage{
strr_process_multi(daily, multiplier = 2, quiet = FALSE)
}
\arguments{
\item{daily}{An unprocessed daily table in the raw AirDNA format, with either
ten or six fields.}

\item{multiplier}{An integer scalar. What multiple of the number of available
processes should the data be combined into?}

\item{quiet}{A logical scalar. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
A processed multilisting table, ready for compression with
\code{\link{strr_compress}}.
}
\description{
\code{strr_process_multi} takes processed daily tables and produces summary
tables listings host activity per day.
}
\details{
A function for cleaning raw daily activity tables from AirDNA and preparing
them for compression into the UPGo format. The function also produces error
files which identify possible corrupt or missing lines in the input file.
}
