% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_process_review.R
\name{strr_process_review}
\alias{strr_process_review}
\title{Process raw STR review tables into UPGo format}
\usage{
strr_process_review(
  review,
  property,
  latest_user = NULL,
  max_id = 0,
  quiet = FALSE
)
}
\arguments{
\item{review}{TKTK}

\item{property}{An strr_property data frame.}

\item{latest_user}{A review_user table with a single row containing the most
recent observations for each user_ID.}

\item{max_id}{The highest review_ID currently in use. New review_IDs will be
assigned starting with max_id + 1.}

\item{quiet}{A logical scalar. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
TKTK.
}
\description{
\code{strr_process_review} takes raw review tables from AirDNA and cleans
them for analysis or for upload in the UPGo database storage format.
}
\details{
A function for cleaning raw review tables from AirDNA and preparing them
for subsequent analysis or upload in the UPGo format.
}
