% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_wrappers.R
\name{with_progress}
\alias{with_progress}
\title{Wrapper for progressr::with_progress}
\usage{
with_progress(expr)
}
\arguments{
\item{expr}{An expression to be evaluated.}
}
\value{
The result of evaluating the expression.
}
\description{
Wrapper for progressr::with_progress
}
