% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_wrappers.R
\name{with_progress2}
\alias{with_progress2}
\title{\code{with_progress} wraps progressr::with_progress if {progressr} is
installed, and returns the original expression if not.}
\usage{
with_progress2(expr)
}
\arguments{
\item{expr}{An expression to be evaluated.}
}
\value{
The result of evaluating the expression.
}
\description{
\code{with_progress} wraps progressr::with_progress if {progressr} is
installed, and returns the original expression if not.
}
