% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_ghost.R
\name{ghost_combine}
\alias{ghost_combine}
\title{Helper function to create potential intersection combinations}
\usage{
ghost_combine(buffers, predicates, n)
}
\arguments{
\item{buffers}{A data frame of buffers from points$data.}

\item{predicates}{The points$data field generated from st_intersects.}

\item{n}{A numeric scalar. The number of points to attempt to find a set of
combinations for}
}
\value{
The output will be a matrix of possible intersection combinations.
}
\description{
\code{ghost_combine} takes `buffers` and `predicates` and generates
combinations.
}
\details{
A function for generating potential intersection combinations using `buffers`
and `predicates` from points$data.
}
