% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_ghost.R
\name{ghost_empty}
\alias{ghost_empty}
\title{Helper function to output an empty ghost hostel table}
\usage{
ghost_empty(points, crs_points)
}
\arguments{
\item{points}{An sf data frame of STR listings nested by cluster.}

\item{crs_points}{The CRS of the points table.}
}
\description{
\code{ghost_empty} produces an empty output table.
}
\details{
A function for producing an empty output table if there are no valid ghost
hostels.
}
