% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_ghost.R
\name{ghost_identify_leftovers}
\alias{ghost_identify_leftovers}
\title{Helper function to identify leftover ghost hotel candidates}
\usage{
ghost_identify_leftovers(points, property_ID, host_ID, min_listings)
}
\arguments{
\item{points}{An sf data frame of STR listings nested by cluster.}

\item{property_ID}{The name of a character or numeric variable in the points
object which uniquely identifies STR listings.}

\item{host_ID}{The name of a character or numeric variable in the points
object which uniquely identifies STR hosts.}

\item{min_listings}{A numeric scalar. The minimum number of listings to
be considered a ghost hotel.}
}
\value{
The output will be a set of new candidate `points`.
}
\description{
\code{ghost_identify_leftovers} identifies leftover ghost hotel candidates.
}
\details{
A function for identifying leftover ghost hotel candidates after the main
ghost hotel analysis has been run.
}
