% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_ghost.R
\name{ghost_intersect_leftovers}
\alias{ghost_intersect_leftovers}
\title{Helper function to rerun analysis on leftover ghost hostel candidates}
\usage{
ghost_intersect_leftovers(property, distance, min_listings)
}
\arguments{
\item{property}{An sf data frame of STR listings nested by cluster.}

\item{distance}{A numeric scalar. The radius (in the units of the CRS) of the
buffer which will be drawn around property to determine possible ghost hostel
locations.}

\item{min_listings}{A numeric scalar. The minimum number of listings to
be considered a ghost hostel.}
}
\value{
The `property` file with additional ghost hostels added.
}
\description{
Helper function to rerun analysis on leftover ghost hostel candidates
}
