% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_ghost_old_helpers.R
\name{ghost_intersect_with_done}
\alias{ghost_intersect_with_done}
\title{Helper function to intersect two buffers}
\usage{
ghost_intersect_with_done(x, y)
}
\arguments{
\item{x}{A buffer.}

\item{y}{A buffer.}
}
\value{
An intersect polygon.
}
\description{
Helper function to intersect two buffers
}
