% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_ghost.R
\name{ghost_stepwise_intersect}
\alias{ghost_stepwise_intersect}
\title{Helper function to intersect a data frame of buffers}
\usage{
ghost_stepwise_intersect(buffers, min_listings)
}
\arguments{
\item{buffers}{A data frame of buffers from property$data.}

\item{min_listings}{A numeric scalar. The minimum number of listings to
be considered a ghost hostel.}
}
\value{
An intersect polygon.
}
\description{
Helper function to intersect a data frame of buffers
}
