% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_compress.R
\name{helper_compress_host}
\alias{helper_compress_host}
\title{Helper function to compress host file}
\usage{
helper_compress_host(data)
}
\arguments{
\item{data}{The processed host table generated through the strr_compress
function.}
}
\value{
The output will be a compressed host table.
}
\description{
\code{helper_compress_host} takes a processed host table and generates a
compressed version.
}
\details{
A helper function for compressing the processed host summary table.
}
