% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_raffle.R
\name{helper_integrate}
\alias{helper_integrate}
\title{Helper function to integrate intersects}
\usage{
helper_integrate(result, pdf, quiet)
}
\arguments{
\item{result, pdf, quiet}{Arguments passed along from the main function.}
}
\value{
The output will be a data frame.
}
\description{
\code{helper_integrate} integrates intersects
}
