% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_raffle.R
\name{helper_intersect}
\alias{helper_intersect}
\title{Helper function to intersect property and polys}
\usage{
helper_intersect(property, polys, distance, quiet)
}
\arguments{
\item{property, polys, distance, quiet}{Arguments passed along from the main
function.}
}
\value{
The output will be a list of two data frames.
}
\description{
\code{helper_intersect} intersects property and polys
}
