% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_helpers.R
\name{helper_progress_message}
\alias{helper_progress_message}
\title{Helper function to display a progress message}
\usage{
helper_progress_message(..., .quiet, .final = FALSE)
}
\arguments{
\item{...}{Character strings to be displayed. The strings can include
code for evaluation via \code{glue::glue} inside `{}`.}

\item{.quiet}{A logical scalar. Should the function execute quietly, or
should it return status updates throughout the function (default)?}

\item{.final}{A logical scalar. Is this the final progress message, to be
formatted in cyan and bold with no time displayed?}
}
\value{
A status message.
}
\description{
\code{helper_progress_message} produces a formatted progress message,
optionally with the current time.
}
