% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_compress.R
\name{strr_compress}
\alias{strr_compress}
\title{Function to compress daily STR tables into UPGo DB format}
\usage{
strr_compress(data, multiplier = 2, quiet = FALSE)
}
\arguments{
\item{data}{A daily table in either the processed UPGo daily format or the
processed UPGo multilisting format.}

\item{multiplier}{An integer scalar. What multiple of the number of available
processes should the data be combined into?}

\item{quiet}{A logical scalar. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
A compressed daily table, ready for upload to a remote database.
}
\description{
\code{strr_compress} takes daily tables (either monthly daily tables from
AirDNA or ML summary tables produced by UPGo) and compresses them into the
UPGo database storage format.
}
\details{
A function for compressing daily activity tables. It takes either AirDNA
daily tables or UPGo multilisting summary tables which have been processed
using \code{\link{strr_process_daily}} or \code{\link{strr_process_multi}},
and converts them into a more storage-efficient one-activity-block-per-row
format.

The output can subsequently be restored to a non-compressed format using
\code{\link{strr_expand}}.
}
