% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_compress.R
\name{strr_compress_helper}
\alias{strr_compress_helper}
\title{Helper function to compress daily file}
\usage{
strr_compress_helper(.data)
}
\arguments{
\item{.data}{The processed daily table generated through the strr_compress
function.}
}
\value{
The output will be a compressed daily table.
}
\description{
\code{strr_compress_helper} takes a processed `daily` table and generates a
compressed version.
}
\details{
A helper function for compressing the processed monthly `daily` table.
}
