% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_expand.R
\name{strr_expand}
\alias{strr_expand}
\title{Expand compressed STR tables}
\usage{
strr_expand(data, quiet = FALSE)
}
\arguments{
\item{data}{A table in compressed UPGo DB format (e.g. created by running
\code{\link{strr_compress}}). Currently daily activity tables and host tables
are recognized.}

\item{quiet}{A logical scalar. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
A table with one row per date and all other fields returned
unaltered.
}
\description{
\code{strr_expand} takes an STR file compressed in the UPGo DB format and
expands it to a one-row-per-date format.
}
\details{
A function for expanding compressed daily activity tables from AirDNA or the
ML summary tables UPGo produces, optionally truncating the table by a
supplied date range. The function can take advantage of multiprocess and/or
remote computation options if a plan is set with the \code{future} package.
}
