% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_expand.R
\name{strr_expand}
\alias{strr_expand}
\title{Function to expand compressed STR tables}
\usage{
strr_expand(.data, start = NULL, end = NULL, cores = 1)
}
\arguments{
\item{.data}{A table in compressed UPGo DB format (e.g. created by running
\code{\link{strr_compress}}). Currently daily activity files and ML daily
summary tables are recognized.}

\item{start}{A character string of format YYYY-MM-DD indicating the
first date to be provided in the output table. If NULL (default), the
earliest date present in the data will be used.}

\item{end}{A character string of format YYYY-MM-DD indicating the
last date to be provided in the output table. If NULL (default), the
latest date present in the data will be used.}

\item{cores}{A positive integer scalar. How many processing cores should be
used to perform the computationally intensive numeric integration step?}
}
\value{
A table with one row per date and all other fields returned unaltered.
}
\description{
\code{strr_expand} takes an STR file compressed in the UPGo DB format and
expands it to a one-row-per-date format.
}
\details{
A function for expanding compressed daily activity tables from AirDNA or the
ML summary tables UPGo produces. The function will also optionally truncate
the table by a supplied date range.
}
