% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_ghost.R
\name{strr_ghost}
\alias{strr_ghost}
\title{Function to identify STR ghost hotels}
\usage{
strr_ghost(points, start_date, end_date, distance = 200,
  min_listings = 3, cores = 1)
}
\arguments{
\item{points}{An sf or sp point-geometry object, in a projected coordinate
system.}

\item{start_date}{A character string of format YYYY-MM-DD indicating the
first date for which to run the analysis.}

\item{end_date}{A character string of format YYYY-MM-DD indicating the last
date for which to run the analysis.}

\item{distance}{A numeric scalar. The radius (in the units of the CRS) of the
buffer which will be drawn around points to determine possible ghost hotel
locations.}

\item{min_listings}{A numeric scalar. The minimum number of listings to
be considered a ghost hotel.}

\item{cores}{A positive integer scalar. How many processing cores should be
used to perform the computationally intensive numeric integration step?}
}
\value{
The output will be a tidy ad dataframe of identified ghost hotels,
  organized with the following fields: `Ghost_ID`: an identifier for each
  unique ghost hotel cluster. `Date`: the date on which the ghost hotel was
  detected. `Host_ID` (or whatever name was passed to the Host_ID argument):
  The ID number of the host operating the ghost hotel. `Listing count`: how
  many separate listings comprised the ghost hotel. `Housing units`: an
  estimate of how many housing units the ghost hotel occupies, calculated as
  `ceiling(Listing_count / 4)`. `Property_IDs` (or whatever name was passed
  to the Property_ID argument): A list of the Property_ID values from the
  listings comprising the ghost hotel. `data`: a nested tibble of additional
  variables present in the points object. `geometry`: the polygons
  representing the possible locations of each ghost hotel.
}
\description{
\code{strr_ghost} takes reported STR listing locations and identifies
possible "ghost hotels"--clusters of private-room STR listings operating in a
single building.
}
\details{
A function for probablistically assigning STR listings to administrative
geographies (e.g. census tracts) based on reported latitude/longitude.
The function works by combining a known probability density function (e.g.
Airbnb's spatial obfuscation of listing locations) with an additional source
of information about possible listing locations--either population or housing
densities.
}
