% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_multi.R
\name{strr_multi}
\alias{strr_multi}
\title{Function to create daily multilisting tables}
\usage{
strr_multi(daily, quiet = FALSE)
}
\arguments{
\item{daily}{An processed daily table in the UPGo format (probably produced
with \code{\link{strr_process_daily}}), with either ten or six fields.}

\item{quiet}{A logical scalar. Should the function execute quietly, or
should it return status updates throughout the function (default)?}
}
\value{
A processed multilisting table, ready for compression with
\code{\link{strr_compress}}.
}
\description{
\code{strr_multi} takes processed daily tables and produces summary
tables listings host activity per day.
}
\details{
A function for aggregating daily activity tables by host, listing type, and
housing status, in order to determine the extent of multilisting activity
in a given STR dataset. This function will typically be run on the output of
\code{\link{strr_process_daily}} in preparation for compression with
\code{\link{strr_compress}}.
}
