% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strr_process_review.R
\name{strr_process_review}
\alias{strr_process_review}
\title{Process raw STR review tables into UPGo format}
\usage{
strr_process_review(review, property, latest_user, max_id = 0, quiet = FALSE)
}
\arguments{
\item{review}{An unprocessed property data frame in the raw AirDNA format,
with either 36 or 56 fields.}

\item{property}{TKTK}

\item{latest_user}{TKTK}

\item{max_id}{The highest review_ID currently in use. New review_IDs will be
assigned starting with max_id + 1.}

\item{quiet}{A logical scalar. Should the function execute quietly, or should
it return status updates throughout the function (default)?}
}
\value{
A list with three elements: 1) the processed property table, with 35
fields; 2) an error table identifying corrupt or otherwise invalid row
entries; 3) a missing_geography table identifying property_IDs with missing
latitude/longitude coordinates.
}
\description{
\code{strr_process_review} takes raw review tables from AirDNA and cleans
them for analysis or for upload in the UPGo database storage format.
}
\details{
A function for cleaning raw review tables from AirDNA and preparing them
for subsequent analysis or upload in the UPGo format.
}
